requirements.txt
flask==1.1.4
flask-restx==0.4.0
flask-cors==3.0.10
flask-caching==1.10.1
flask-sqlalchemy==2.5.1
flask-jwt-extended==4.3.1
flask-brcrypt==1.0.1
pyyaml==6.0
markupsafe==2.0.1
jsonschema==4.4.0
boto3==1.22.7
celery==4.4.6